Add llvm 12.0.0-rc4
Also move gcc from base to dev: clang, lld and the runtime libraries in
this package replace gcc as the system compiler for the distribution.

This llvm package can bootstrap from a gcc system image. Run

  pkg make llvm llvm-stage1

to build a stage-1 compiler, linker and libraries. These still contain
links to gcc libatomic and libgcc_s, so overwrite the system image with
the llvm-stage1 image then build again with

  pkg make llvm llvm-stage2

The result is a fully independent llvm toolchain and runtime which can
be used to rebuild the rest of the system.

When bootstrapping, llvm uses $CHOST as its host triple, which pkg gets
from the existing toolchain with  'cc -dumpmachine'. This may need to be
overridden if it isn't already 'x86_64-arachsys-linux-musl' or
'aarch64-arachsys-linux-musl'.