Simplify llvm build recipe
Use the LLVM_ENABLE_RUNTIMES mechanism to simplify our llvm build.

This doesn't work for compiler-rt as the sanitizers depend on libunwind
and libc++abi, but LLVM_ENABLE_PROJECTS can be used for compiler-rt and
its sanitizers while other libraries are built via LLVM_ENABLE_RUNTIMES.

The new structure works for bootstrapping too: on a gcc system with
libstdc++ instead of libc++, the sanitizers are disabled to avoid link
failures against libc++abi and libunwind before these are available.