Introduce multi-image package builds
Allow package builds to generate multiple images in /dst/img/* as an
alternative to a single image in /dst.

Pass through the package name from the main build command to the helper
so it can cleanly handle the special-case of a single image in /dst, and
in turn offer both package name and features/options as arguments to the
prepare and build functions.

Fix the error messages from the tidy function in lib/shared to include
full paths to illegal files, which is helpful in the multi-image case.