Introduce image requirements
An optional file 'requirements' at the top level of an image records a
list of runtime dependencies, by analogy with the require directives in
package builds declaring build dependencies. Refactor the pkg tool to
handle these, treating naked files in $IMGDIR as meta-images: empty
images with requirements listed in the file.

One such meta-image is 'base' generated by package 'meta'. This captures
a list of core images implicitly required at the start of every build.
It doesn't make sense to try to deduce this from the contents of the
package category 'base' anymore, as multi-image packages break the
one-to-one correspondence between package recipes and image names.

To make encapsulated builds more robust, run them in a temporary mount
namespace using a helper to set up the directory hierarchy and bind
mounts. This allows read-write access to the build root from the host at
the same time as a protected read-only mount for the build job itself.