Add rust beta-20210407
At the time of packaging, stable rust is incompatible with llvm 12.0.0, but
when the current beta branch is released, this recipe will switch to track
stable releases.

Building rust requires a recent pre-existing rust compiler and cargo. When
bootstrapping from upstream binary linux-musl builds, the gcc libraries and
a symlink from /lib/ld-musl-$(arch).so.1 to ld.so need to be added to the
build chroot. Replace "--enable-local-rust --local-rust-root=" with
"--set=target.$(arch)-unknown-linux-musl.crt-static=false" in the configure
arguments to use upstream musl binaries but with dynamic-linking enabled.