Tidy up packaging of python
Python 3.x hard-codes a lot of false assumptions about paths, which require
invasive patching. It is particularly bad at dealing with an install in
/bin and /lib: an empty prefix is confused with a relative path whereas a
prefix of '/' isn't handled carefully enough to avoid references to //bin
and //lib/python.

Move the library directory from /lib/python3.9 to /lib/python.

Install pip as part of the system python image but with the annoying nag to
upgrade to the latest version disabled.