Add openssl 3.0.0-beta1
Although OpenSSL is still a mess with a crazy Perl build system and reams
of terrifying cruft, LibreSSL is nonetheless falling behind. For example,
Ed25519 is still missing in libcrypto and the OpenSSL 1.1 TLSv1.3 API is
still not supported by libssl.

Python is likely to drop LibreSSL support entirely in 3.10 unless this
situation improves, and it isn't clear if OpenBSD will maintain patches
to keep it working there.

Hopefully LibreSSL will catch up sufficiently that we are not forced to
switch to an uglier library, but in case not, this package is preparation
in advance.