Introduce pkg check to query for upstream updates
If a package PKG has a file PKG/check, pkg check will use it check for
upstream updates. Example contents are

  url https://musl.libc.org/releases/
  match 'musl-([0-9.]+)\.tar' '\1'
  version 1.2.2

The url command defines a URL which is retrieved and searched for substrings
matching an extended regular expression given as the first argument of the
match command. For each match, the second argument is used to construct an
available version number, typically using backreferences to the pattern. The
most recent available version found (in the sense of sort -V) is compared
against the packaged version which is declared with the version command.

If match is given only one argument, the second argument defaults to '\1'.

Invoked with a list of packages, pkg check will check each of them. Without
arguments, it will check all packages in $PKGDIR.