Fail with an explicit pkg error when / isn't a mount point
On systems where mounts are shared, unshare recursively resets their
propagation to private in the new mount namespace. Although this is default
behaviour, request it explicitly as we rely on it for correct operation.

Drop --make-[r]slave when bind mounting in the namespace as all mounts are
already private.

If the root directory isn't a mount point, unshare fails because it can't
recursively set the propagation from the root. This happens when chrooted
into a directory which is not itself a (bind) mount. Detect and report this
case explicitly to avoid a more obscure error message from unshare.