Fix an obscure race in parallel pkg check
Individual check jobs are launched from a loop in a shell pipeline. Add a
wait at the end of this loop to prevent the loop subprocess from exiting
before its background children complete.

We rely on $PPID in the internal pkg contain helper to test if the mount
namespace has been unshared from that of the parent process. If the parent
exits early, we will be reparented to init and instead compare against its
mount namespace. This gives a false positive if we're running nested or in
another (partial) container without an enclosing PID namespace.