Rename linux-headers to linux
With pkg build, this package continues to generate the linux-headers image
for use with musl. However, it is also useful with pkg prepare to generate
patched kernel sources ready to compile.

To this end, add our local kernel patches for building with llvm and clang,
for ncurses/musl compatibility of menuconfig, and to suppress an annoying
nl80211 warning about self-managed regdoms with iwd.

For example,

  pkg prepare ~/linux
  zcat /proc/config.gz > ~/linux/.config
  pkg shell -s ~/linux bc bison flex perl
  make -j $(nproc) -C /src

might now be used to build a kernel with pkg encapsulation.