Introduce pkg system for constructing bootstrap system images
In a similar way to pkg shell, this constructs a system container image
but the image is copied into a specified directory or used to create a tar
archive instead of being discarded on exit.

We copy the whole package recipe tree and the images for included packages
into root's home directory in the new system, ready for bootstrapping.

The contents of the system bin/ and lib/ directories are hard-linked to
the corresponding contents in home/root/img/*/ to avoid doubling the size
of the image.