Update llvm to cmake 3.21.4
Although cmake 3.22.0 has been released, building llvm with this fails when
installing libc++.so, as this is a linker script rather than a symlink. The
build system tries to use file RPATH_CHANGE to remove any build-time rpath,
which unsurprisingly generates an error as the script is not an ELF file.