Improve pkg support for multi-image recipes
Expose an image command to package build functions which can be used to
direct installed content in /dst to specific images. Files not otherwise
allocated are sent to an image with the same name as the package recipe.

Rework the image tidy function to process multiple images at once for a
cleaner build implementation in the helper action.