Distinguish between 'base' and 'build' image lists
The build infrastructure including llvm, clang and make is very large
compared to the rest of the base images, and it isn't needed in many
system images. Separate the list of images required to build with pkg
from the true base list.

Add init, less and procps to the base system. Their footprint is small
and they are invariably needed for a usable working image.