Drop openssl package
This recipe was reluctantly added in preparation for an enforced migration
if LibreSSL couldn't catch up with the OpenSSL 1.1 API, with major users
like Python dropping compatibility. Happily, the situation is much improved
with the 3.4.x releases of LibreSSL, so we can avoid subjecting ourselves
to OpenSSL's weird Perl build system and legacy cruft.

(Ironically, an experimental global rebuild with OpenSSL 3.0 actually had
more compatibility problems than LibreSSL as a libcrypto/libssl provider.)