Rework package build process and metadata representation
Move the requirements list to a top-level info subdirectory in each image
directory. Introduce a file ('source') naming the package recipe used to
build the image.

Extend the image directive in recipe build functions to allow both content
and requirements to be declared for each image that is generated. Update
package recipes to use this.

Change the build-time image target directory from /dst/img to /img for
cleaner separation between installation and packaging areas.