Build gcc with itself unless we're bootstrapping
Retain isl.diff although it's only needed when building with clang, to
ensure the bootstrap build continues to work fine.

Drop the workaround for a binutils installation bug that is now fixed.

Only check for minor version updates of binutils. Shared-tree builds need
the corresponding binutils and gcc release branches to be contemporaneous.