Teach kakoune to fall back to the 256-colour palette
Add a terminal_ui option 'terminal_use_palette' which instructs kakoune to
approximate 24-bit colours using the standard 256-colour palette instead
of emitting true RGB escape sequences, which not all terminals support.

The palette contains 24 scales of grey as well as the 6x6x6 RGB colour
cube, so use the former for better brightness resolution if the quantised
colour coordinates lie on the cube diagonal.

Strictly, it would be more correct to pick the closest palette entry in
a perceptually-linear colour space like CIELAB. Back in RGB space, the
effect of this is to distort the previously planar boundaries between
adjacent quantised colours. However, the benefit of this is very small
unless the palette is also remapped from a cartesian RGB grid, and the
calculation would be expensive to perform on the fly while generating
terminal escapes from faces.