Evolve pkg contain into pkg sandbox
Use a PID namespace in addition to the mount namespace to better isolate
the build container. Only populate /dev with the minimum required subset
of safe nodes rather than all the host devices. We can drop the /sys mount
completely but /proc is still required for ps; it may be possible to mount
with subset=pid in future when procps supports this better.

When unshare supports /etc/subuid and /etc/subgid with --map-auto, wrap
unprivileged invocations of pkg sandbox in a user namespace, allowing a
non-root user to build packages.