Fix the llvm libunwind.so linker script
When pkg switched /dst to be a symlink to /img/PKGNAME, the llvm recipe
broke on non-x86-64 architectures, where libunwind.so linker script is
required. Fix the recipe to generate the relative path correctly again.

Instead of special-casing x86-64, always install the linker script for
libunwind.so, so we catch problems with it earlier and more reliably.