Fix paths in python's bundled wheels
The ensurepip module is used to install pip and setuptools when creating
a venv. It bundles wheels for both of these. Although the wheels claim to
be architecture and OS independent, they actually hard-code incorrect
assumptions about the include and library paths.

Unpack, patch and repack these binary archives when building python. This
is untidy, but it not really possible to do better given the embarrassing
state of packaging upstream.