Simplify pkg system
Drop support for output in tar format and the -t TEMPLATE option which is
only useful with archive output.

Allow the target directory to be specified as '-', in which case the image
will be built in a temporary directory inside $IMGDIR, the path to which is
then returned on standard output.