Improve local openntpd patches
Use adjtimex rather than adjtime for adjusting the clock as musl fails
with EINVAL instead of saturating when the offset is too large.

Add a -s option to ntpd to specify that it should exit with an error
instead of daemonising if it fails to set the initial time at startup.