Teach openssh to respect HOME
Upstream ssh, ssh-add and ssh-keygen always consult getpwuid(getuid()) to
find the user's home directory. Teach them to prefer the value of HOME in
the environment if set. This ensures sane behaviour of ssh inside fakeroot
or pseudo-style user namespaces.

There is no security impact as an attacker with the ability to manipulate
the environment can already use LD_PRELOAD to override getpwuid(), given
that none of these commands is setuid.