Update grep to version 3.8
This release of grep links to pcre2 by default so remove the old pcre
library from the base image list. (Nothing else in the distribution
links to it, so there are no manual require directives to add.)

As the old egrep and fgrep now emit obsolescence warnings, drop them
completely so we can find and clear up remaining users.