Run pkg build in an isolated network namespace
We need to download sources during pkg prepare but the subsequent pkg
build should never access the network. Enforce this by running it in an
isolated network namespace.

Set HOME as /tmp rather than / in the sandbox so it is writeable by
default. Because the prepare and build phases each get a fresh (empty)
/tmp, using it helps ensure that any dependencies and state are properly
included in the /src tree hash.

Fix the recipes that manually set HOME to /tmp to assume this default.