Fix dnsmasq DHCPv6 behaviour with no IPv6 prefixes
If DHCPv6 is enabled but no prefixes are available, reply with NoAddrsAvail
rather than ignoring the request and logging a noisy warning. This allows
DHCPv6 to be enabled unconditionally in ra-stateless mode, doing the right
thing when IPv6 prefixes come and go at runtime.

Drop the ULA router advertisement patch as this is no longer needed and is
not suitable for upstream. It breaks configurations where multiple local
prefixes are routed between sites or NPTv6 is used to translate static ULA
addresses to dynamic global ones on the router.