Drop helix
This is an expensive build for a relatively unused package, with a 825MB
source tree bundling lots of unaudited third-party tree-sitter grammars
of distinctly variable quality. A further 640MB of dependencies are pulled
in on top of this by cargo.

It is interesting to compare this vast 1.5GB source tree with the 11MB to
build kakoune, 10MB to build joe, 18MB to build micro, 14MB to build nano,
70MB to build vim, etc.