Accept a pattern for pkg check version match
Allow check recipes to include wildcard specifications such as

  check() {
    url https://upstream.org/releases
    match version '1.3.*' 'v([0-9.]+)\.tar'
  }

This is occasionally handy for chopping a continuous release into
milestones for packaging when there are no canonical ones upstream.

Take this opportunity to check for unintentional pattern interpolation
on the RHS of [[ x == y ]] and [[ x != y ]] expressions throughout the
pkg codebase, quoting where necessary.