Remove vestigial code for enable docs
We always strip man pages and other documentation from packages during
the tidy phase of the build process. Although theoretically this can be
disabled by enabling the 'docs' feature on the pkg build command line,
almost every recipe now assumes it will happen. Often documentation build
dependencies aren't met or are patched out, and install steps are skipped.

Drop the remnants of enable docs from pkg and from the coreutils recipe,
which was the only package that tried to respect it.