Move unshared code from pkg/lib/shared to pkg/actions/*
In particular, move tidy() into the body of the build helper, its only
call site. This enables some simplification because most operations can
run over /img as a whole rather than each subdirectory in turn. Similarly,
define tree() in the body of the prepare helper, although the underlying
hashtree() remains in the shared library.

For the prepare and build helpers, source and run the recipe in a subshell
to reduce the chance of it interfering with the subsequent check and tidy
operations within the sandbox.

Strip special files, absolute symlinks and dangling symlinks from prepared
source trees, although some of these are also caught when calculating
a tree checksum if one is declared in the recipe. We needn't check for
setuid or setgid binaries as the relevant bits will be cleared while
normalising file ownership and permissions.