Drop explicit --with-lto from python configure
The --enable-optimizations flag turns on all the stable optimisations
for both gcc and llvm, but forcing --with-lto on gcc builds gives a 35M
larger static libpython and a 4x slower build time as the gcc toolchain
doesn't support -flto=thin.

Increase THREAD_STACK_SIZE to 2MB to match Alpine Linux's value, which
is apparently necessary for some pathological situations.