Fix some problems with util-linux sulogin
When sulogin is run on /dev/console and there are multiple console devices,
it is smart enough to fork and prompt for login on all of them. However,
the -t option is then broken: sulogin returns as soon as any prompt times
out, even when a shell has been launched from a different prompt.

Fix this behaviour, waiting for any shells to exit before allowing the
parent sulogin to exit. Use the special exit code 142 (corresponding to
SIGALRM) when no logins were made before timeout; otherwise return the
exit status of the last shell to terminate. Ctrl-D at the prompt still
forces an immediate timeout.

With a short timeout, sulogin is then suitable for an initramfs /init
script, as well as early failure cases in /etc/init without a timeout.