Don't daemonize openntpd if it is already session leader
On Arachsys Linux, openntpd is typically run as

  daemon await-net ntpd -p /run/ntpd.pid

to ensure the network is up so the clock can be set from https constraints.

In this case, ntpd is already in the background so calling daemon()
just unnecessarily forks and messes up the process hierarchy. If ntpd is
already session leader on invocation, don't call daemon().