Update firecracker to version 1.4.1 again
Upstream publicly tagged a 1.4.1 release yesterday, then deleted the
tag after a couple of hours, and finally retagged a later commit as
v1.4.1 today. Some context for this debacle is in

  https://github.com/firecracker-microvm/firecracker/pull/4013

Cached downloads of

  https://github.com/firecracker-microvm/firecracker/archive/refs/tags/v1.4.1.tar.gz

will need to be manually deleted, and any clones of the repository that
pulled during the window where the first tag existed will retain the
incorrect v1.4.1 tag unless it is manually removed and pulled again.