Patch musl to continue host search after NODATA as well as NXDOMAIN
When a search list is provided in /etc/resolv.conf, musl will search each
given domain for unqualified hosts, proceeding to the next candidate if
it receives an NXDOMAIN response, then finally looking up the bare name.

If a name has no direct A or AAAA records but there are other records
on or beneath it, nameservers should return NODATA instead of NXDOMAIN.
In this case, musl aborts the search with an EAI_NODATA error instead
of continuing.

Whilst this is a principled interpretation of the RFCs, much of DNS's
protocol design is breathtakingly incompetent. Musl's behaviour interacts
poorly with mainstream strategies to work around the worst of this, such
as CloudFlare's 'black lies' used to generate signed negative responses
on-the-fly without expensive global knowledge of a zone:

  https://datatracker.ietf.org/doc/html/draft-valsorda-dnsop-black-lies

Even wildcard records of non-host-related types such as TXT or MX are
sufficient to cause problems with this search strategy.

Other resolvers including glibc, FreeBSD's libc and Go's DNS client
continue searching after NODATA. Adjust musl to match everyone else.