Update llvm to version 17.0.1
Link /bin/as to clang not llvm-as: llvm-as is an llvm bytecode assembler
not a host architecture assembler and shouldn't be installed as part of
the toolchain.

Provide the correct paths to liblzma.so, libz.so and libzstd.so.
It's fine for llvm to link against these as they are in base. Admire in
passing the incompetently inconsistent naming of LibEdit_INCLUDE_DIRS,
LIBLZMA_INCLUDE_DIR, ZLIB_INCLUDE_DIR and zstd_INCLUDE_DIR.

Stop trying to patch around the LLVM_ENABLE_PER_TARGET_RUNDIR_DIR
installation behaviour to get both builtins and host libraries in the
right place. Fix up any paths incorrectly containing /$CHOST/ after make
install has finished. This is more robust and easier to forward-port than
an intrusive patch to the tangled cmake build infrastructure.