Stop linux dropping proxy neighbours on link flap
Recent kernels flush the neighbour cache when a link carrier goes down, as
it could be a sign of the port changing networks. Unfortunately, they also
carelessly drop proxy entries at the same time. Correct this behaviour.

Static (permanent) configuration shouldn't be lost on link flaps in
any case, but it's also logically incorrect to flush proxy entries on
carrier drop. Proxy entries specify what the host should advertise TO an
attached network, not what is reachable FROM that network.