Fix broken tree checksum for firecracker again
Upstream publicly tagged a 1.5.0 release last week, then later silently
updated the tag. The 1.4.1 release in August was botched in a similar way.

This broke our tree checksum, but more generally, cached downloads of

  https://github.com/firecracker-microvm/firecracker/archive/refs/tags/v1.5.0.tar.gz

will need to be manually deleted, and any clones of the repository that
pulled during the window where the first tag existed will retain the
incorrect v1.5.0 tag unless it is manually removed and pulled again.