Include debug info in musl and llvm static runtimes
We strip binaries and shared libraries for smaller runtime system images,
but debug symbols are nonetheless useful for development. Leave them in
place in static libraries. They can still be stripped when statically
linking a production binary if necessary.

Many static libraries were already built with symbols, but the musl and
llvm recipes need adjusting to ensure they are included.