Drop local direct IO patch for firecracker
This is quite a lot of work to forward port when deck chairs are being
randomly shuffled in affected areas of the codebase. The original upstream
pull request

  https://github.com/firecracker-microvm/firecracker/pull/3236

was rejected. Retaining it would be a permanent maintenance burden when
packaging new releases.

Despite the crazy complexity and dependencies that are unconditionally
bundled to support the ugly and pointless mmds overlay, avoiding
double-buffering for more efficient block device access apparently lacks
a clear use case.