Add prctl PR_SET_NO_NEW_PRIVS support to openssh chroot
When ChrootDirectory is used in sshd_config, sshd descends the given
path to check each component can be written only by root. This protects
setuid/setgid binaries from being run in untrusted roots, e.g. with
user-controlled dynamic linker or libraries.

A linux-specific alternative is to use prctl PR_SET_NO_NEW_PRIVS to disable
setuid and setgid altogether for the chrooted process. This is especially
convenient for granting sftp-only logins to subdirectories of a user's home
directory with internal-sftp, which would otherwise be clumsy to implement.

Prefix the ChrootDirectory path with safe: to block setuid/setgid
using this prctl. Otherwise, setuid/setgid will be unrestricted and the
traditional ownership checks will be made.