Add a clang workaround for musl's sched.h
The musl sched.h definitions for CPU_SET_S(), CPU_CLR_S() and CPU_ISSET_S()
warn when passed a long cpu argument instead of an int with -Wsign-compare
on clang. Squelch this by ensuring a long is promoted from signed to
unsigned in the same way as an int is.

This fixes the only compiler warning when building the python package.