Set default CARGO_HOME to ~/.cache/cargo
If CARGO_HOME is not set, upstream cargo dumps its cache in ~/.cargo rather
than respecting the XDG structure. Change this default to ~/.cache/cargo.

This isn't a perfect fix as it doesn't respect the XDG_* variables and
CARGO_HOME should really be split between ~/.config, ~/.local and ~/.cache.
However, given protracted upstream indolence and prevarication on this
issue, an incomplete solution is better than nothing.