Update python to version 3.13.0
Don't require the tests used for profile-guided optimisation to succeed,
partially reverting upstream commit 6ab60400. We have a number of
uninteresting failures because of tests which expect hard-coded paths that
we change or assume non-utf8 locale support. The only genuine failure is
that fma(tiny, -tiny, 0.0) == 0.0 instead of -0.0, which is a bug fixed
in upstream musl commit 9683bd62.

No longer drop the dbm module as dbm.sqlite gives it a backend we support.