Ensure rust patches cover all vendored crate versions
In a majestic demonstration of the dependency culture cargo engenders,
rust 1.81.0 manages to vendor three versions of curl-sys, two versions
of home, three versions of openssl and four versions of openssl-sys.
Make sure curl.diff, home.diff and libressl.diff patch all of these rather
than just one or two versions of each, fixing several subtle bugs in the
resulting cargo binary.

Add a reminder to the build recipe that we need to re-check every version
is properly patched when upgrading to a new rust release.